Claw lcm.regime behind the construction perimeter
Extends the beartype claw to `lcm.regime` with `INTERNAL_CONF`. The
explicit `@beartype(conf=REGIME_CONF)` decorator on the `Regime` and
`MarkovTransition` constructors still wins, so construction-time type
violations keep surfacing as `RegimeInitializationError`.

`_default_H` receives state/value arrays whose dtype follows user
input — a `discount_factor` supplied as `1` arrives as an int32 array,
as `0.95` a float32 array — so its parameters are annotated `NumericND`
(a new `FloatND | IntND` alias in `lcm.typing`). The named alias keeps
`dags.tree`'s params-template extraction printing `NumericND` rather
than a bare `Union`. The `_IdentityTransition.__call__` discrete test
now passes an int32 array, matching the `DiscreteState` contract the
claw enforces.

+++ tests/test_regime.py
@@ -193,8 +193,8 @@ def test_identity_transition_call():
 def test_identity_transition_discrete():
     """Identity transition works for discrete states."""
     identity = _IdentityTransition("education", annotation=DiscreteState)
-    result = identity(education=jnp.array(1))
-    assert result == jnp.array(1)
+    result = identity(education=jnp.array(1, dtype=jnp.int32))
+    assert result == jnp.array(1, dtype=jnp.int32)
 
 
 def test_identity_transition_name():
